clean up contracts
