+++ fluent/mod.ts
@@ -1,3 +1,4 @@
+export * from "./AddressRune.ts"
 export * from "./ClientRune.ts"
 export * from "./CodecRune.ts"
 export * from "./ConstRune.ts"
@@ -7,10 +8,10 @@ export * from "./ExtrinsicStatusRune.ts"
 export * from "./FinalizedBlockRune.ts"
 export * from "./HeaderRune.ts"
 export * from "./InBlockBlockRune.ts"
-export * from "./InkContract.ts"
 export * from "./MetadataRune.ts"
 export * from "./MultisigRune.ts"
 export * from "./PalletRune.ts"
+export * from "./PublicKeyRune.ts"
 export * from "./rpc_method_runes.ts"
 export * from "./rpc_runes.ts"
 export * from "./SignedExtrinsicRune.ts"
