--- ink_metadata/mod.ts
@@ -1,3 +1,0 @@
-export * from "./known.ts"
-export * from "./Metadata.ts"
-export * from "./normalize.ts"
