+++ mod.ts
@@ -3,7 +3,7 @@ export { BitSequence } from "./deps/scale.ts"
 export * as Z from "./deps/zones.ts"
 export * from "./fluent/mod.ts"
 export * as frame from "./frame_metadata/mod.ts"
-export * as ink from "./ink_metadata/mod.ts"
+export * as ink from "./ink/mod.ts"
 export * from "./primitives/mod.ts"
 export * as rpc from "./rpc/mod.ts"
 export * from "./rune/mod.ts"
